requirements.txt
python-dotenv==1.0.1
python_fasthtml==0.10.0
pytz==2024.2
supabase==2.10.0
